Moved FrameAdapter classes to separate files
